[nemo-qml-plugin-notifications] Deprecate origin property. Contributes to JB#39479
a) there doesn't seem to be real-life use for this, and
b) it's not too well defined what it really means. Best guess
what it could be used for is some sort of notification grouping
based on unique identifier for sender.

Sailfish has had kind of avatar support, but it's not really an avatar
sending notifications and there are plenty of image/icon properties
that fit such purpose better.