[qml-plugin-notifications] Remove bad click example. Contributes to JB#51116
There is no default action by default so this part of the example didn't
work. It's also debatable if minimal example should include that at
all so just removing now.

Was considering also alternative of adding the "default" by default
but that would change the current sailfish behavior which requires
no actions on banner style. Could change the rules for that, but
adding the default here would prevent any differentiation between
notifications with actions and those without.