[notifications] Support passing action type info. JB#55557
Implicitly use "input" if there's input info set for the action.
The type is easier to handle on the notification manager side
and the existing way to pass the "input" map verbatim from app
to d-bus feels suspicious.

Think we should revisit the whole action text parameter details, but
at the moment it's not too much exposed to the apps nor supported
so we can get back into that later.